lambda_function/requirements.txt
pillow>=10.2.0
openai>=1.14.3